Data.ByteString.Lazy

Description

A time and space-efficient implementation of lazy byte vectors using lists of packed Word8 arrays, suitable for high performance use, both in terms of large data quantities, or high speed requirements. Lazy ByteStrings are encoded as lazy lists of strict chunks of bytes.

A key feature of lazy ByteStrings is the means to manipulate large or unbounded streams of data without requiring the entire sequence to be resident in memory. To take advantage of this you have to write your functions in a lazy streaming style, e.g. classic pipeline composition. The default I/O chunk size is 32k, which should be good in most circumstances.

Some operations, such as concat, append, reverse and cons, have better complexity than their Data.ByteString equivalents, due to optimisations resulting from the list spine structure. For other operations lazy ByteStrings are usually within a few percent of strict ones.

The recomended way to assemble lazy ByteStrings from smaller parts is to use the builder monoid from Data.ByteString.Builder.

This module is intended to be imported qualified, to avoid name clashes with Prelude functions. eg.

import qualified Data.ByteString.Lazy as B

Original GHC implementation by Bryan O'Sullivan. Rewritten to use UArray by Simon Marlow. Rewritten to support slices and use ForeignPtr by David Roundy. Rewritten again and extended by Don Stewart and Duncan Coutts. Lazy variant by Duncan Coutts and Don Stewart.

Lazy ByteString

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A lazy ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Lazy.Char8 it can be interpreted as containing 8-bit characters.

type LazyByteString = ByteString

Type synonym for the lazy flavour of ByteString.

Since: bytestring-0.11.2.0

Introducing and eliminating ByteStrings

empty :: ByteString

O(1) The empty ByteString

singleton :: Word8 -> ByteString

O(1) Convert a Word8 into a ByteString

pack :: [Word8] -> ByteString

O(n) Convert a '[Word8]' into a ByteString.

unpack :: ByteString -> [Word8]

O(n) Converts a ByteString to a '[Word8]'.

fromStrict :: ByteString -> ByteString

O(1) Convert a strict ByteString into a lazy ByteString.

toStrict :: ByteString -> ByteString

O(n) Convert a lazy ByteString into a strict ByteString.

Note that this is an expensive operation that forces the whole lazy ByteString into memory and then copies all the data. If possible, try to avoid converting back and forth between strict and lazy bytestrings.

fromChunks :: [ByteString] -> ByteString

O(c) Convert a list of strict ByteString into a lazy ByteString

toChunks :: ByteString -> [ByteString]

O(c) Convert a lazy ByteString into a list of strict ByteString

foldrChunks :: (ByteString -> a -> a) -> a -> ByteString -> a

Consume the chunks of a lazy ByteString with a natural right fold.

foldlChunks :: (a -> ByteString -> a) -> a -> ByteString -> a

Consume the chunks of a lazy ByteString with a strict, tail-recursive, accumulating left fold.

Basic interface

cons :: Word8 -> ByteString -> ByteString infixr 5

O(1) cons is analogous to (:) for lists.

cons' :: Word8 -> ByteString -> ByteString infixr 5

O(1) Unlike cons, cons' is strict in the ByteString that we are consing onto. More precisely, it forces the head and the first chunk. It does this because, for space efficiency, it may coalesce the new byte onto the first 'chunk' rather than starting a new 'chunk'.

So that means you can't use a lazy recursive contruction like this:

let xs = cons' c xs in xs

You can however use cons, as well as repeat and cycle, to build infinite lazy ByteStrings.

snoc :: ByteString -> Word8 -> ByteString infixl 5

O(n/c) Append a byte to the end of a ByteString

append :: ByteString -> ByteString -> ByteString

O(n/c) Append two ByteStrings

head :: HasCallStack => ByteString -> Word8

O(1) Extract the first element of a ByteString, which must be non-empty.

This is a partial function, consider using uncons instead.

uncons :: ByteString -> Maybe (Word8, ByteString)

O(1) Extract the head and tail of a ByteString, returning Nothing if it is empty.

unsnoc :: ByteString -> Maybe (ByteString, Word8)

O(n/c) Extract the init and last of a ByteString, returning Nothing if it is empty.

It is no faster than using init and last

last :: HasCallStack => ByteString -> Word8

O(n/c) Extract the last element of a ByteString, which must be finite and non-empty.

This is a partial function, consider using unsnoc instead.

tail :: HasCallStack => ByteString -> ByteString

O(1) Extract the elements after the head of a ByteString, which must be non-empty.

This is a partial function, consider using uncons instead.

init :: HasCallStack => ByteString -> ByteString

O(n/c) Returns all the elements of a ByteString except the last one.

This is a partial function, consider using unsnoc instead.

null :: ByteString -> Bool

O(1) Test whether a ByteString is empty.

length :: ByteString -> Int64

O(c) length returns the length of a ByteString as an Int64

Transforming ByteStrings

map :: (Word8 -> Word8) -> ByteString -> ByteString

O(n) map f xs is the ByteString obtained by applying f to each element of xs.

reverse :: ByteString -> ByteString

O(n) reverse xs returns the elements of xs in reverse order.

intersperse :: Word8 -> ByteString -> ByteString

The intersperse function takes a Word8 and a ByteString and `intersperses' that byte between the elements of the ByteString. It is analogous to the intersperse function on Lists.

intercalate :: ByteString -> [ByteString] -> ByteString

O(n) The intercalate function takes a ByteString and a list of ByteStrings and concatenates the list after interspersing the first argument between each element of the list.

transpose :: [ByteString] -> [ByteString]

The transpose function transposes the rows and columns of its ByteString argument.

Reducing ByteStrings (folds)

foldl :: (a -> Word8 -> a) -> a -> ByteString -> a

foldl, applied to a binary operator, a starting value (typically the left-identity of the operator), and a ByteString, reduces the ByteString using the binary operator, from left to right.

foldl' :: (a -> Word8 -> a) -> a -> ByteString -> a

foldl' is like foldl, but strict in the accumulator.

foldl1 :: HasCallStack => (Word8 -> Word8 -> Word8) -> ByteString -> Word8

foldl1 is a variant of foldl that has no starting value argument, and thus must be applied to non-empty ByteStrings.

foldl1' :: HasCallStack => (Word8 -> Word8 -> Word8) -> ByteString -> Word8

foldl1' is like foldl1, but strict in the accumulator.

foldr :: (Word8 -> a -> a) -> a -> ByteString -> a

foldr, applied to a binary operator, a starting value (typically the right-identity of the operator), and a ByteString, reduces the ByteString using the binary operator, from right to left.

foldr' :: (Word8 -> a -> a) -> a -> ByteString -> a

foldr' is like foldr, but strict in the accumulator.

Since: bytestring-0.11.2.0

foldr1 :: HasCallStack => (Word8 -> Word8 -> Word8) -> ByteString -> Word8

foldr1 is a variant of foldr that has no starting value argument, and thus must be applied to non-empty ByteStrings

foldr1' :: HasCallStack => (Word8 -> Word8 -> Word8) -> ByteString -> Word8

foldr1' is like foldr1, but strict in the accumulator.

Since: bytestring-0.11.2.0

Special folds

concat :: [ByteString] -> ByteString

O(n) Concatenate a list of ByteStrings.

concatMap :: (Word8 -> ByteString) -> ByteString -> ByteString

Map a function over a ByteString and concatenate the results

any :: (Word8 -> Bool) -> ByteString -> Bool

O(n) Applied to a predicate and a ByteString, any determines if any element of the ByteString satisfies the predicate.

all :: (Word8 -> Bool) -> ByteString -> Bool

O(n) Applied to a predicate and a ByteString, all determines if all elements of the ByteString satisfy the predicate.

maximum :: HasCallStack => ByteString -> Word8

O(n) maximum returns the maximum value from a ByteString

minimum :: HasCallStack => ByteString -> Word8

O(n) minimum returns the minimum value from a ByteString

compareLength :: ByteString -> Int64 -> Ordering

O(c) compareLength compares the length of a ByteString to an Int64

Since: bytestring-0.11.1.0

Building ByteStrings

Scans

scanl

Arguments

:: (Word8 -> Word8 -> Word8)
accumulator -> element -> new accumulator

-> Word8
starting value of accumulator

-> ByteString
input of length n

-> ByteString
output of length n+1

scanl is similar to foldl, but returns a list of successive reduced values from the left.

scanl f z [x1, x2, ...] == [z, z `f` x1, (z `f` x1) `f` x2, ...]

Note that

head (scanl f z xs) == z
last (scanl f z xs) == foldl f z xs

scanl1 :: (Word8 -> Word8 -> Word8) -> ByteString -> ByteString

scanl1 is a variant of scanl that has no starting value argument.

scanl1 f [x1, x2, ...] == [x1, x1 `f` x2, ...]

Since: bytestring-0.11.2.0

scanr

Arguments

:: (Word8 -> Word8 -> Word8)
element -> accumulator -> new accumulator

-> Word8
starting value of accumulator

-> ByteString
input of length n

-> ByteString
output of length n+1

scanr is similar to foldr, but returns a list of successive reduced values from the right.

scanr f z [..., x{n-1}, xn] == [..., x{n-1} `f` (xn `f` z), xn `f` z, z]

Note that

head (scanr f z xs) == foldr f z xs
last (scanr f z xs) == z

Since: bytestring-0.11.2.0

scanr1 :: (Word8 -> Word8 -> Word8) -> ByteString -> ByteString

scanr1 is a variant of scanr that has no starting value argument.

Since: bytestring-0.11.2.0

Accumulating maps

mapAccumL :: (acc -> Word8 -> (acc, Word8)) -> acc -> ByteString -> (acc, ByteString)

The mapAccumL function behaves like a combination of map and foldl; it applies a function to each element of a ByteString, passing an accumulating parameter from left to right, and returning a final value of this accumulator together with the new ByteString.

mapAccumR :: (acc -> Word8 -> (acc, Word8)) -> acc -> ByteString -> (acc, ByteString)

The mapAccumR function behaves like a combination of map and foldr; it applies a function to each element of a ByteString, passing an accumulating parameter from right to left, and returning a final value of this accumulator together with the new ByteString.

Infinite ByteStrings

repeat :: Word8 -> ByteString

repeat x is an infinite ByteString, with x the value of every element.

replicate :: Int64 -> Word8 -> ByteString

O(n) replicate n x is a ByteString of length n with x the value of every element.

cycle :: HasCallStack => ByteString -> ByteString

cycle ties a finite ByteString into a circular one, or equivalently, the infinite repetition of the original ByteString.

iterate :: (Word8 -> Word8) -> Word8 -> ByteString

iterate f x returns an infinite ByteString of repeated applications of f to x:

iterate f x == [x, f x, f (f x), ...]

Unfolding ByteStrings

unfoldr :: (a -> Maybe (Word8, a)) -> a -> ByteString

O(n) The unfoldr function is analogous to the List 'unfoldr'. unfoldr builds a ByteString from a seed value. The function takes the element and returns Nothing if it is done producing the ByteString or returns Just (a,b), in which case, a is a prepending to the ByteString and b is used as the next element in a recursive call.

Substrings

Breaking strings

take :: Int64 -> ByteString -> ByteString

O(n/c) take n, applied to a ByteString xs, returns the prefix of xs of length n, or xs itself if n > length xs.

takeEnd :: Int64 -> ByteString -> ByteString

O(c) takeEnd n xs is equivalent to drop (length xs - n) xs. Takes n elements from end of bytestring.

>>> takeEnd 3 "abcdefg"
"efg"
>>> takeEnd 0 "abcdefg"
""
>>> takeEnd 4 "abc"
"abc"

Since: bytestring-0.11.2.0

drop :: Int64 -> ByteString -> ByteString

O(n/c) drop n xs returns the suffix of xs after the first n elements, or empty if n > length xs.

dropEnd :: Int64 -> ByteString -> ByteString

O(n) dropEnd n xs is equivalent to take (length xs - n) xs. Drops n elements from end of bytestring.

>>> dropEnd 3 "abcdefg"
"abcd"
>>> dropEnd 0 "abcdefg"
"abcdefg"
>>> dropEnd 4 "abc"
""

Since: bytestring-0.11.2.0

splitAt :: Int64 -> ByteString -> (ByteString, ByteString)

O(n/c) splitAt n xs is equivalent to (take n xs, drop n xs).

takeWhile :: (Word8 -> Bool) -> ByteString -> ByteString

Similar to takeWhile, returns the longest (possibly empty) prefix of elements satisfying the predicate.

takeWhileEnd :: (Word8 -> Bool) -> ByteString -> ByteString

Returns the longest (possibly empty) suffix of elements satisfying the predicate.

takeWhileEnd p is equivalent to reverse . takeWhile p . reverse.

>>> {-# LANGUAGE OverloadedLists #-)
>>> takeWhileEnd even [1,2,3,4,6]
[4,6]

Since: bytestring-0.11.2.0

dropWhile :: (Word8 -> Bool) -> ByteString -> ByteString

Similar to dropWhile, drops the longest (possibly empty) prefix of elements satisfying the predicate and returns the remainder.

dropWhileEnd :: (Word8 -> Bool) -> ByteString -> ByteString

Similar to dropWhileEnd, drops the longest (possibly empty) suffix of elements satisfying the predicate and returns the remainder.

dropWhileEnd p is equivalent to reverse . dropWhile p . reverse.

>>> {-# LANGUAGE OverloadedLists #-)
>>> dropWhileEnd even [1,2,3,4,6]
[1,2,3]

Since: bytestring-0.11.2.0

span :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)

Similar to span, returns the longest (possibly empty) prefix of elements satisfying the predicate and the remainder of the string.

span p is equivalent to break (not . p) and to (takeWhile p &&& dropWhile p).

spanEnd :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)

Returns the longest (possibly empty) suffix of elements satisfying the predicate and the remainder of the string.

spanEnd p is equivalent to breakEnd (not . p) and to (takeWhileEnd p &&& dropWhileEnd p).

We have

spanEnd (not . isSpace) "x y z" == ("x y ", "z")

and

spanEnd (not . isSpace) ps
   ==
let (x, y) = span (not . isSpace) (reverse ps) in (reverse y, reverse x)

Since: bytestring-0.11.2.0

break :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)

Similar to break, returns the longest (possibly empty) prefix of elements which do not satisfy the predicate and the remainder of the string.

break p is equivalent to span (not . p) and to (takeWhile (not . p) &&& dropWhile (not . p)).

breakEnd :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)

Returns the longest (possibly empty) suffix of elements which do not satisfy the predicate and the remainder of the string.

breakEnd p is equivalent to spanEnd (not . p) and to (takeWhileEnd (not . p) &&& dropWhileEnd (not . p)).

Since: bytestring-0.11.2.0

group :: ByteString -> [ByteString]

The group function takes a ByteString and returns a list of ByteStrings such that the concatenation of the result is equal to the argument. Moreover, each string in the result contains only equal elements. For example,

group "Mississippi" = ["M","i","ss","i","ss","i","pp","i"]

It is a special case of groupBy, which allows the programmer to supply their own equality test.

groupBy :: (Word8 -> Word8 -> Bool) -> ByteString -> [ByteString]

The groupBy function is the non-overloaded version of group.

inits :: ByteString -> [ByteString]

Returns all initial segments of the given ByteString, shortest first.

tails :: ByteString -> [ByteString]

O(n) Returns all final segments of the given ByteString, longest first.

initsNE :: ByteString -> NonEmpty ByteString

Returns all initial segments of the given ByteString, shortest first.

Since: bytestring-0.11.4.0

tailsNE :: ByteString -> NonEmpty ByteString

O(n) Returns all final segments of the given ByteString, longest first.

Since: bytestring-0.11.4.0

stripPrefix :: ByteString -> ByteString -> Maybe ByteString

O(n) The stripPrefix function takes two ByteStrings and returns Just the remainder of the second iff the first is its prefix, and otherwise Nothing.

Since: bytestring-0.10.8.0

stripSuffix :: ByteString -> ByteString -> Maybe ByteString

O(n) The stripSuffix function takes two ByteStrings and returns Just the remainder of the second iff the first is its suffix, and otherwise Nothing.

Breaking into many substrings

split :: Word8 -> ByteString -> [ByteString]

O(n) Break a ByteString into pieces separated by the byte argument, consuming the delimiter. I.e.

split 10  "a\nb\nd\ne" == ["a","b","d","e"]   -- fromEnum '\n' == 10
split 97  "aXaXaXa"    == ["","X","X","X",""] -- fromEnum 'a' == 97
split 120 "x"          == ["",""]             -- fromEnum 'x' == 120
split undefined ""     == []                  -- and not [""]

and

intercalate [c] . split c == id
split == splitWith . (==)

As for all splitting functions in this library, this function does not copy the substrings, it just constructs new ByteStrings that are slices of the original.

splitWith :: (Word8 -> Bool) -> ByteString -> [ByteString]

O(n) Splits a ByteString into components delimited by separators, where the predicate returns True for a separator element. The resulting components do not contain the separators. Two adjacent separators result in an empty component in the output. eg.

splitWith (==97) "aabbaca" == ["","","bb","c",""] -- fromEnum 'a' == 97
splitWith undefined ""     == []                  -- and not [""]

Predicates

isPrefixOf :: ByteString -> ByteString -> Bool

O(n) The isPrefixOf function takes two ByteStrings and returns True iff the first is a prefix of the second.

isSuffixOf :: ByteString -> ByteString -> Bool

O(n) The isSuffixOf function takes two ByteStrings and returns True iff the first is a suffix of the second.

The following holds:

isSuffixOf x y == reverse x `isPrefixOf` reverse y

Search for arbitrary substrings

Searching ByteStrings

Searching by equality

elem :: Word8 -> ByteString -> Bool

O(n) elem is the ByteString membership predicate.

notElem :: Word8 -> ByteString -> Bool

O(n) notElem is the inverse of elem

Searching with a predicate

find :: (Word8 -> Bool) -> ByteString -> Maybe Word8

O(n) The find function takes a predicate and a ByteString, and returns the first element in matching the predicate, or Nothing if there is no such element.

find f p = case findIndex f p of Just n -> Just (p ! n) ; _ -> Nothing

filter :: (Word8 -> Bool) -> ByteString -> ByteString

O(n) filter, applied to a predicate and a ByteString, returns a ByteString containing those characters that satisfy the predicate.

partition :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)

O(n) The partition function takes a predicate a ByteString and returns the pair of ByteStrings with elements which do and do not satisfy the predicate, respectively; i.e.,

partition p bs == (filter p xs, filter (not . p) xs)

Indexing ByteStrings

index :: HasCallStack => ByteString -> Int64 -> Word8

O(c) ByteString index (subscript) operator, starting from 0.

This is a partial function, consider using indexMaybe instead.

indexMaybe :: ByteString -> Int64 -> Maybe Word8

O(c) ByteString index, starting from 0, that returns Just if:

0 <= n < length bs

Since: bytestring-0.11.0.0

(!?) :: ByteString -> Int64 -> Maybe Word8

O(1) ByteString index, starting from 0, that returns Just if:

0 <= n < length bs

Since: bytestring-0.11.0.0

elemIndex :: Word8 -> ByteString -> Maybe Int64

O(n) The elemIndex function returns the index of the first element in the given ByteString which is equal to the query element, or Nothing if there is no such element. This implementation uses memchr(3).

elemIndexEnd :: Word8 -> ByteString -> Maybe Int64

O(n) The elemIndexEnd function returns the last index of the element in the given ByteString which is equal to the query element, or Nothing if there is no such element. The following holds:

elemIndexEnd c xs = case elemIndex c (reverse xs) of
  Nothing -> Nothing
  Just i  -> Just (length xs - 1 - i)

Since: bytestring-0.10.6.0

elemIndices :: Word8 -> ByteString -> [Int64]

O(n) The elemIndices function extends elemIndex, by returning the indices of all elements equal to the query element, in ascending order. This implementation uses memchr(3).

findIndex :: (Word8 -> Bool) -> ByteString -> Maybe Int64

The findIndex function takes a predicate and a ByteString and returns the index of the first element in the ByteString satisfying the predicate.

findIndexEnd :: (Word8 -> Bool) -> ByteString -> Maybe Int64

The findIndexEnd function takes a predicate and a ByteString and returns the index of the last element in the ByteString satisfying the predicate.

Since: bytestring-0.10.12.0

findIndices :: (Word8 -> Bool) -> ByteString -> [Int64]

The findIndices function extends findIndex, by returning the indices of all elements satisfying the predicate, in ascending order.

count :: Word8 -> ByteString -> Int64

count returns the number of times its argument appears in the ByteString

count = length . elemIndices

But more efficiently than using length on the intermediate list.

Zipping and unzipping ByteStrings

zip :: ByteString -> ByteString -> [(Word8, Word8)]

O(n) zip takes two ByteStrings and returns a list of corresponding pairs of bytes. If one input ByteString is short, excess elements of the longer ByteString are discarded. This is equivalent to a pair of unpack operations.

zipWith :: (Word8 -> Word8 -> a) -> ByteString -> ByteString -> [a]

zipWith generalises zip by zipping with the function given as the first argument, instead of a tupling function. For example, zipWith (+) is applied to two ByteStrings to produce the list of corresponding sums.

packZipWith :: (Word8 -> Word8 -> Word8) -> ByteString -> ByteString -> ByteString

A specialised version of zipWith for the common case of a simultaneous map over two ByteStrings, to build a 3rd.

Since: bytestring-0.11.1.0

unzip :: [(Word8, Word8)] -> (ByteString, ByteString)

O(n) unzip transforms a list of pairs of bytes into a pair of ByteStrings. Note that this performs two pack operations.

Ordered ByteStrings

Low level conversions

Copying ByteStrings

copy :: ByteString -> ByteString

O(n) Make a copy of the ByteString with its own storage. This is mainly useful to allow the rest of the data pointed to by the ByteString to be garbage collected, for example if a large string has been read in, and only a small part of it is needed in the rest of the program.

I/O with ByteStrings

⚠ Using lazy I/O functions like readFile or hGetContents means that the order of operations such as closing the file handle is left at the discretion of the RTS. Hence, the developer can face some issues when:

The program reads a file and writes the same file. This means that the file may be locked because the handler has not been released when writeFile is executed.

The program reads thousands of files, but due to lazy evaluation, the OS's file descriptor limit is reached before the handlers can be released.

Why?

Consider the following program:

import qualified Data.ByteString.Lazy as BL
main = do
  _ <- BL.readFile "foo.txt"
  BL.writeFile "foo.txt" mempty

Generally, in the IO monad side effects happen sequentially and in full. Therefore, one might reasonably expect that reading the whole file via readFile executes all three actions (open the file handle, read its content, close the file handle) before control moves to the following writeFile action. This expectation holds for the strict Data.ByteString API. However, the above lazy ByteString variant of the program fails with openBinaryFile: resource busy (file is locked).

The reason for this is that Data.ByteString.Lazy is specifically designed to handle large or unbounded streams of data incrementally, without requiring all the data to be resident in memory at the same time. Incremental processing would not be possible if readFile were to follow the usual rules of IO: evaluating all side effects would require reading the file in full and closing its handle before returning from readFile. This is why readFile (and hGetContents in general) is implemented via unsafeInterleaveIO, which allows IO side effects to be delayed and interleaved with subsequent processing of the return value. That's exactly what happens in the example above: readFile opens a file handle, but since the content is not fully consumed, the file handle remains open, allowing the content to read on demand (never in this case, since the return value is ignored). So when writeFile is executed next, foo.txt is still open for reading and the RTS takes care to avoid simultaneously opening it for writing, instead returning the error shown above.

How to enforce the order of effects?

If the content is small enough to fit in memory, consider using strict readFile, potentially applying fromStrict afterwards. E. g.,

import qualified Data.ByteString as BS
import qualified Data.ByteString.Lazy as BL
main = do
  _ <- BS.readFile "foo.txt"
  BL.writeFile "foo.txt" mempty

If you are dealing with large or unbounded data streams, consider reaching out for a specialised package, such as conduit, machines-bytestring, pipes-bytestring, streaming-bytestring, streamly-bytestring, etc.

Standard input and output

getContents :: IO ByteString

getContents. Equivalent to hGetContents stdin. Will read lazily

putStr :: ByteString -> IO ()

Write a ByteString to stdout.

The chunks will be written one at a time. Other threads might write to the stdout in between, and hence putStr alone is not suitable for concurrent writes.

interact :: (ByteString -> ByteString) -> IO ()

The interact function takes a function of type ByteString -> ByteString as its argument. The entire input from the standard input device is passed to this function as its argument, and the resulting string is output on the standard output device.

Files

readFile :: FilePath -> IO ByteString

Read an entire file lazily into a ByteString.

The Handle will be held open until EOF is encountered.

Note that this function's implementation relies on hGetContents. The reader is advised to read its documentation.

writeFile :: FilePath -> ByteString -> IO ()

Write a ByteString to a file.

appendFile :: FilePath -> ByteString -> IO ()

Append a ByteString to a file.

I/O with Handles

hGetContents :: Handle -> IO ByteString

Read entire handle contents lazily into a ByteString. Chunks are read on demand, using the default chunk size.

File handles are closed on EOF if all the file is read, or through garbage collection otherwise.

hGet :: Handle -> Int -> IO ByteString

Read n bytes into a ByteString, directly from the specified Handle.

hGetNonBlocking :: Handle -> Int -> IO ByteString

hGetNonBlocking is similar to hGet, except that it will never block waiting for data to become available, instead it returns only whatever data is available. If there is no data available to be read, hGetNonBlocking returns empty.

Note: on Windows and with Haskell implementation other than GHC, this function does not work correctly; it behaves identically to hGet.

hPut :: Handle -> ByteString -> IO ()

Outputs a ByteString to the specified Handle.

The chunks will be written one at a time. Other threads might write to the Handle in between, and hence hPut alone is not suitable for concurrent writes.

hPutNonBlocking :: Handle -> ByteString -> IO ByteString

Similar to hPut except that it will never block. Instead it returns any tail that did not get written. This tail may be empty in the case that the whole string was written, or the whole original string if nothing was written. Partial writes are also possible.

Note: on Windows and with Haskell implementation other than GHC, this function does not work correctly; it behaves identically to hPut.

hPutStr :: Handle -> ByteString -> IO ()

A synonym for hPut, for compatibility