Data.ByteString.Lazy.Builder.ASCII

Description

Deprecated: Use Data.ByteString.Builder instead

We decided to rename the Builder modules. Sorry about that.

In additon, the ASCII module has been merged into the main Data.ByteString.Builder module.

The old names will hang about for at least once release cycle and then later remove them.

Documentation

module Data.ByteString.Builder

byteStringHexFixed :: ByteString -> Builder

lazyByteStringHexFixed :: ByteString -> Builder