Data.ByteString.Lazy.Builder.Extras

Description

We decided to rename the Builder modules. Sorry about that.

The old names will hang about for at least once release cycle before we deprecate them and then later remove them.

Documentation

module Data.ByteString.Builder.Extra