Data.ByteString.Lazy.Internal

Description

A module containing semi-public ByteString internals. This exposes the ByteString representation and low level construction functions. Modules which extend the ByteString system will need to use this module while ideally most users will be able to make do with the public interface modules.

The lazy ByteString type and representation

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A lazy ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Lazy.Char8 it can be interpreted as containing 8-bit characters.

Constructors

Empty

Chunk !ByteString ByteString

type LazyByteString = ByteString

Type synonym for the lazy flavour of ByteString.

Since: bytestring-0.11.2.0

chunk :: ByteString -> ByteString -> ByteString

Smart constructor for Chunk. Guarantees the data type invariant.

foldrChunks :: (ByteString -> a -> a) -> a -> ByteString -> a

Consume the chunks of a lazy ByteString with a natural right fold.

foldlChunks :: (a -> ByteString -> a) -> a -> ByteString -> a

Consume the chunks of a lazy ByteString with a strict, tail-recursive, accumulating left fold.

Data type invariant and abstraction function

invariant :: ByteString -> Bool

The data type invariant: Every ByteString is either Empty or consists of non-null StrictByteStrings. All functions must preserve this.

checkInvariant :: ByteString -> ByteString

Lazily checks that the given ByteString satisfies the data type's "no empty chunks" invariant, raising an exception in place of the first chunk that does not satisfy the invariant.

Chunk allocation sizes

defaultChunkSize :: Int

The chunk size used for I/O. Currently set to 32k, less the memory management overhead

smallChunkSize :: Int

The recommended chunk size. Currently set to 4k, less the memory management overhead

chunkOverhead :: Int

The memory management overhead. Currently this is tuned for GHC only.

Conversion with lists: packing and unpacking

packBytes :: [Word8] -> ByteString

packChars :: [Char] -> ByteString

unpackBytes :: ByteString -> [Word8]

unpackChars :: ByteString -> [Char]

Conversions with strict ByteString

fromStrict :: ByteString -> ByteString

O(1) Convert a strict ByteString into a lazy ByteString.

toStrict :: ByteString -> ByteString

O(n) Convert a lazy ByteString into a strict ByteString.

Note that this is an expensive operation that forces the whole lazy ByteString into memory and then copies all the data. If possible, try to avoid converting back and forth between strict and lazy bytestrings.