Data.ByteString.Short.Internal

Description

Internal representation of ShortByteString

The ShortByteString type and representation

data ShortByteString

A compact representation of a Word8 vector.

It has a lower memory overhead than a ByteString and does not contribute to heap fragmentation. It can be converted to or from a ByteString (at the cost of copying the string data). It supports very few other operations.

It is suitable for use as an internal representation for code that needs to keep many short strings in memory, but it should not be used as an interchange type. That is, it should not generally be used in public APIs. The ByteString type is usually more suitable for use in interfaces; it is more flexible and it supports a wide range of operations.

Constructors

SBS ByteArray#

Conversions

toShort :: ByteString -> ShortByteString

O(n). Convert a ByteString into a ShortByteString.

This makes a copy, so does not retain the input string.

fromShort :: ShortByteString -> ByteString

O(n). Convert a ShortByteString into a ByteString.

pack :: [Word8] -> ShortByteString

O(n). Convert a list into a ShortByteString

unpack :: ShortByteString -> [Word8]

O(n). Convert a ShortByteString into a list.

Other operations

empty :: ShortByteString

O(1). The empty ShortByteString.

null :: ShortByteString -> Bool

O(1) Test whether a ShortByteString is empty.

length :: ShortByteString -> Int

O(1) The length of a ShortByteString.

index :: ShortByteString -> Int -> Word8

O(1) ShortByteString index (subscript) operator, starting from 0.

unsafeIndex :: ShortByteString -> Int -> Word8

Low level operations

createFromPtr

Arguments

:: Ptr a
source data

-> Int
number of bytes to copy

-> IO ShortByteString

copyToPtr

Arguments

:: ShortByteString
source data

-> Int
offset into source

-> Ptr a
destination

-> Int
number of bytes to copy

-> IO ()

Low level conversions

Packing CStrings and pointers

packCString :: CString -> IO ShortByteString

O(n). Construct a new ShortByteString from a CString. The resulting ShortByteString is an immutable copy of the original CString, and is managed on the Haskell heap. The original CString must be null terminated.

Since: bytestring-0.10.10.0

packCStringLen :: CStringLen -> IO ShortByteString

O(n). Construct a new ShortByteString from a CStringLen. The resulting ShortByteString is an immutable copy of the original CStringLen. The ShortByteString is a normal Haskell value and will be managed on the Haskell heap.

Since: bytestring-0.10.10.0

Using ByteStrings as CStrings

useAsCString :: ShortByteString -> (CString -> IO a) -> IO a

O(n) construction. Use a ShortByteString with a function requiring a null-terminated CString. The CString is a copy and will be freed automatically; it must not be stored or used after the subcomputation finishes.

Since: bytestring-0.10.10.0

useAsCStringLen :: ShortByteString -> (CStringLen -> IO a) -> IO a

O(n) construction. Use a ShortByteString with a function requiring a CStringLen. As for useAsCString this function makes a copy of the original ShortByteString. It must not be stored or used after the subcomputation finishes.

Since: bytestring-0.10.10.0